requirements.txt
django-autocomplete-light==3.3.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-extensions==2.1.4
django-filter==2.0.0
django-mass-edit==3.2.0
django-resetdb
django-sql-explorer==1.1.2
django-stronghold==0.3.0
django-tables2==2.0.3
django_import_data
docxtpl==0.5.17
ipython-genutils==0.2.0
ipython==7.2.0
psycopg2-binary==2.7.6.1
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.6
pyexcel==0.5.10
pykml==0.1.3
requests==2.21.0
tablib==0.12.1
tabulate==0.8.2
tqdm==4.30.0
